// src/models/npcs/actors/stats/professions/middle/types.ts
export type MiddleStratum =
  | 'gentry'
  | 'priest'
  | 'sorcerer'
  | 'scholar'
  | 'artisan'
  | 'surgeon'
  | 'merchant'
  | 'ship captain'
  | 'bureaucrat'
  | 'officer (military)'
  | 'criminal (boss)'

export type ArtisanSpecializations =
  | 'alchemist'
  | 'artificer'
  | 'baker'
  | 'blacksmith'
  | 'brewer'
  | 'butcher'
  | 'cobbler'
  | 'distiller'
  | 'fletcher'
  | 'glassblower'
  | 'goldsmith'
  | 'herbalist'
  | 'jeweler'
  | 'leatherworker'
  | 'locksmith'
  | 'shipwright'
  | 'silversmith'
  | 'tailor'
  | 'tanner'
  | 'mechanic'
  | 'vintner'
  | 'weaver'